python-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Background job scheduling
apscheduler==3.10.4

# Database
supabase==2.3.4

# HTTP client for async requests
aiohttp==3.9.1

# Vector similarity search
faiss-cpu==1.7.4
# For GPU acceleration (uncomment if using CUDA)
# faiss-gpu==1.7.4

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80
requests==2.31.0

# ML Models
torch==2.1.2
torchvision==0.16.2

# CLIP (OpenAI's image encoder)
git+https://github.com/openai/CLIP.git

# OCR
paddleocr==2.7.0.3
paddlepaddle==2.6.0

# Face Detection & Emotion (optional - for full ML features)
# retinaface==0.0.17  # Version not available, use 1.1.1 if needed
# fer==22.5.1
# mtcnn==0.1.1

# Text Processing
transformers==4.37.0
sentence-transformers==2.3.1

# Utilities
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
python-dotenv==1.0.0

# YouTube Intelligence Brain dependencies
joblib==1.3.2  # For model serialization
pandas==2.1.4  # For data analysis
matplotlib==3.8.2  # For visualization (optional)
seaborn==0.13.0  # For statistical plots (optional)

# Optional: GPU acceleration (uncomment if using CUDA)
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118

